+++ BulletN.h
@@ -4,7 +4,7 @@
 #define BULLET_N_BOX_HEIGHT 4
 class CBulletN :public CBullet {
 public:
-	CBulletN(float x, float y, int angle, bool friendly);
+	CBulletN(float x, float y, float angle, bool friendly);
 	void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects = NULL);
 	void CreateBox(DWORD dt);
 };
