+++ BulletS.h
@@ -5,7 +5,7 @@
 class CBulletS:public CBullet {
 
 public:
-	CBulletS(float x, float y, int angle, bool friendly);
+	CBulletS(float x, float y, float angle, bool friendly);
 	void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects = NULL);
 	void CreateBox(DWORD dt);
 };
